+++ src/main/java/org/alex/service/AssignmentService.java
@@ -2,6 +2,7 @@ package org.alex.service;
 
 import org.alex.api.service.IAssignmentService;
 import org.alex.entity.Assignment;
+import org.alex.exception.IllegalStringException;
 import org.alex.repository.AssignmentRepository;
 
 import java.util.List;
@@ -12,22 +13,38 @@ public class AssignmentService implements IAssignmentService {
 
     @Override
     public void create(Assignment assignment) {
-        assignmentRepository.create(assignment);
+        if (assignment == null) {
+            throw new NullPointerException();
+        } else {
+            assignmentRepository.create(assignment);
+        }
     }
 
     @Override
-    public void delete(String fromId, String toId) {
-        assignmentRepository.delete(fromId, toId);
+    public void delete(String fromId, String toId) throws IllegalStringException {
+        if (fromId != null && toId != null) {
+            assignmentRepository.delete(fromId, toId);
+        } else {
+            throw new IllegalStringException();
+        }
     }
 
     @Override
     public List<Assignment> getAllById(String fromId) {
-        return assignmentRepository.getAllById(fromId);
+        if (fromId == null || fromId == "") {
+            throw new NullPointerException();
+        } else {
+            return assignmentRepository.getAllById(fromId);
+        }
     }
 
     @Override
     public void merge(List<Assignment> assignments) {
-        assignmentRepository.merge(assignments);
+        if (assignments == null) {
+            throw new NullPointerException();
+        } else {
+            assignmentRepository.merge(assignments);
+        }
     }
 
     @Override
